Adding a continuous build yml file
