+++ .travis.yml
@@ -12,6 +12,11 @@ addons:
       - libboost-dev
 
 before_install:
+  # The lz4 library requires a special repository
+  - sudo add-apt-repository -y ppa:nathan-renniewaldock/ppa
+  - sudo apt-get update
+  - sudo apt-get install liblz4-dev 
+  
   # install updated version of cmake
   - wget http://www.cmake.org/files/v3.3/cmake-3.3.0-rc3-Linux-x86_64.sh
   - sh cmake-3.3.0-rc3-Linux-x86_64.sh --prefix=$HOME --exclude-subdir
@@ -26,6 +31,9 @@ before_install:
   - sudo add-apt-repository -y ppa:boost-latest/ppa
   - sudo apt-get update
   - sudo apt-get install boost1.55
+
+  # install support compression libraries
+  - sudo apt-get install libsnappy-dev libsnappy1
 
 script:
   # Report version of CMake and GCC
