+++ CMakeLists.txt
@@ -82,7 +82,7 @@ add_library(glnexus
             include/genotyper.h src/genotyper.cc
             include/KeyValue.h src/KeyValue.cc
             include/BCFKeyValueData.h src/BCFKeyValueData.cc
-            include/RocksIntf.h src/RocksIntf.cc)
+            include/RocksKeyValue.h src/RocksKeyValue.cc)
 add_dependencies(glnexus htslib)
 add_dependencies(glnexus rocksdb)
 add_dependencies(glnexus yaml-cpp)
@@ -93,11 +93,6 @@ add_dependencies(playground libglnexus)
 target_link_libraries(playground -pthread jemalloc glnexus libhts z librocksdb libyaml-cpp)
 
 install(TARGETS playground DESTINATION bin)
-
-add_executable(rock_toy src/rock_toy.cc)
-target_link_libraries(rock_toy -pthread jemalloc librocksdb snappy z lz4 bz2)
-
-install(TARGETS rock_toy DESTINATION bin)
 
 ################################
 # Testing
@@ -123,11 +118,13 @@ if (test)
                  src/diploid.cc test/diploid.cc
                  test/unit_tests.cc
                  test/htslib_behaviors.cc
+                 test/rocks_behaviors.cc
                  test/types.cc
                  test/service.cc
-                 test/MemDBTest.cc test/RocksDBTest.cc)
+                 test/BCFKeyValueData.cc
+                 test/rocks_integration.cc)
   add_dependencies(unit_tests libglnexus)
-  target_link_libraries(unit_tests glnexus librocksdb -pthread jemalloc libhts rt snappy z bz2 lz4 libyaml-cpp)
+  target_link_libraries(unit_tests glnexus librocksdb -pthread jemalloc libhts boost_system boost_filesystem rt snappy z bz2 lz4 rt libyaml-cpp)
 
   add_test(unit_tests unit_tests)
 endif()
